step 2 - blog model
